Split CLI into loader and mapper.
